Fix newline being omitted in pre blocks
The br tags within pre blocks were being omitted by the default
rule in the converter. This was replaced by a custom rule for pre
tag that collects each line while considering br tags.

This could raise issues where code and other special tags inside
pre blocks may be ignored. This has to be kept an eye on for.